requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Django==5.2.5
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==25.2
django-cors-headers==4.7.0
django-csp==4.0
django-extensions==4.1
django-multiselectfield==1.0.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dlib==20.0.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
MouseInfo==0.1.3
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pdfkit==1.0.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-magic-bin==0.4.14
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
scipy==1.16.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.9.0
